#114: Move 'section' based components into subdirectory
